requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic[email]
web3
eth-account